Use the ocm fork of the kube rbac proxy
In disconnected environments, the previous image might not be present
because it was not specified in the related images of the operator. But
we have a fork of the kube-rbac-proxy that we are maintaining, so this
can be supplied at deploy-time easily.

That image is already mentioned in the relatedImages section of the
operator's packagemanifest, so it seems nothing needs to be added there.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/15904

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>