Remove the 3des cipher used by the kube-rbac-proxy by default
The metrics service endpoint is an httpd connection using the default
go ciphers which includes `TLS_RSA_WITH_3DES_EDE_CBC_SHA`.  This cipher
is graded as a `C` due to sweet32.  Even though the only use of this
connection is for internal metric gathering, we want to move up to
TLS 1.3 which doesn't have that cipher.

Refs:
 - https://github.com/stolostron/backlog/issues/14361

Signed-off-by: Gus Parvin <gparvin@redhat.com>